Fix `PlanarSet` typings
The `add` method was giving a return type of `PlanarSet`, which is
incompatible with `Set#add`. In such cases, the return value should be
specified as `this`.

I have not reviewed the new typings file in detail, but it seems to
work OK at least as regards the parts of the library that I use in my
app.